Refs #3009. Fixing errors creating linux installer

+++ examples/C++/Configurability/UseCasePublisher.cpp
@@ -138,7 +138,7 @@ int main(){
             std::cout << "Please input a valid argument" << std::endl;
             continue;
         }
-        user_configuration.depth = choice;
+        user_configuration.depth = (uint8_t)choice;
         validinput=true;
     }
     std::cout << "---------------------------------------------------" << std::endl;
@@ -155,7 +155,7 @@ int main(){
             continue;
         }
         if(choice > 0){
-        user_configuration.depth = choice;
+        user_configuration.depth = (uint8_t)choice;
         }else{
         user_configuration.depth = 1;
         std::cout << "Defaulting to 1 instance..." << std::endl;
@@ -176,7 +176,7 @@ int main(){
             std::cout << "Please input a valid argument" << std::endl;
             continue;
         }
-        user_configuration.max_samples_per_key = choice;
+        user_configuration.max_samples_per_key = (uint16_t)choice;
         validinput=true;
     }
 
@@ -215,7 +215,7 @@ int main(){
             std::cout << "Please input a valid argument" << std::endl;
             continue;
         }
-        user_configuration.history_size = choice;
+        user_configuration.history_size = (uint16_t)choice;
         validinput=true;
     }
